+++ cmd/conduit/cli/cli.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cli
 
 import (
 	"os"
@@ -23,7 +23,7 @@ import (
 	"github.com/conduitio/ecdysis"
 )
 
-func RunCLI(cfg conduit.Config) {
+func Run(cfg conduit.Config) {
 	e := ecdysis.New(ecdysis.WithDecorators(cecdysis.CommandWithExecuteWithClientDecorator{}))
 
 	cmd := e.MustBuildCobraCommand(&root.RootCommand{DefaultCfg: cfg})
